[Gold I] Title: 골목 대장 호석 - 효율성 2, Time: 296 ms, Memory: 49464 KB, Score: 28 point -BaekjoonHub

+++ 백준/Gold/20183. 골목 대장 호석 － 효율성 2/골목 대장 호석 － 효율성 2.cc
@@ -50,7 +50,7 @@ void solve() {
 			int next = p.first;
 			ll nW = p.second.first;
 			int nShame = p.second.second;
-			int distance = dist[now] + nW;
+			ll distance = dist[now] + nW;
 			if (dist[next] > distance && distance < minDist[next]) {
 				minDist[next] = distance;
 				pq.push({ next, {distance, max(shame, nShame)} });
